Basic AI
Basic AI has been implemented
Still have some bugs (for some reason character jitters when it has reached the correct angle)
Will have to work on getting the AI to attack

+++ Magic World/Assets/Scripts/Battle System/CharacterTurn.cs
@@ -8,9 +8,9 @@ namespace BattleSystem
     public class CharacterTurn
     {
         public CharacterManager character;
-        public float turnTime;
+        public int turnTime;
 
-        public CharacterTurn(CharacterManager newCharacter, float newTurnTime) {
+        public CharacterTurn(CharacterManager newCharacter, int newTurnTime) {
             character = newCharacter;
             turnTime = newTurnTime;
         }
